refactor: 리쿠르트 관련 enum 파일 이동
