+++ src/components/Forms/RecruitForm/Fields/EndDate/index.tsx
@@ -5,10 +5,8 @@ import dayjs from 'dayjs';
 import { useFormState, useWatch } from 'react-hook-form';
 
 import { DatePicker } from '~/components/Common';
-import {
-  RecruitCategoryName,
-  useRecruitFormContext,
-} from '~/components/Forms/RecruitForm/utils';
+import { useRecruitFormContext } from '~/components/Forms/RecruitForm/utils';
+import { RecruitCategoryName } from '~/services/recruit';
 import { flex } from '~/styles/utils';
 
 const fieldName = 'endDate';
