+++ src/services/recruit/utils/types.ts
@@ -18,6 +18,28 @@ export type SkillType =
   | 'IOS'
   | 'Flutter';
 
+export type RecruitType =
+  | '프론트엔드'
+  | '백엔드'
+  | '기획/디자인'
+  | '앱'
+  | '스터디';
+
+export interface LimitType {
+  recruitType: RecruitType;
+  limit: number;
+  currentNumber?: number;
+}
+
+export interface SkillsType {
+  name: SkillName;
+  skillId: number;
+}
+
+export type RecruitCategoryType = 'study' | 'project';
+
+/* enums */
+
 export enum SkillName {
   SPRING = 'Spring',
   REACT = 'React',
@@ -40,26 +62,19 @@ export enum SkillName {
 }
 export const SkillNameSet = new Set<string>(Object.values(SkillName));
 
-export type RecruitType =
-  | '프론트엔드'
-  | '백엔드'
-  | '기획/디자인'
-  | '앱'
-  | '스터디';
-
-export interface LimitType {
-  recruitType: RecruitType;
-  limit: number;
-  currentNumber?: number;
-}
-
-export interface SkillsType {
-  name: SkillName;
-  skillId: number;
-}
-
-export type RecruitCategoryType = 'study' | 'project';
-
 // ---- Parts ----
 
-// ---- Applicants ----
+export enum RecruitParts {
+  FRONTEND = '프론트엔드',
+  BACKEND = '백엔드',
+  PM = '기획/디자인',
+  APP = '앱',
+  STUDY = '스터디',
+}
+
+// ---- Category ----
+
+export enum RecruitCategoryName {
+  PROJECT = '프로젝트',
+  STUDY = '스터디',
+}
